requirement.txt
asgiref==3.2.5
awscli==1.17.4
beautifulsoup4==4.8.1
boto3==1.10.45
botocore==1.14.4
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
colorama==0.4.1
cryptography==2.8
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.4
django-allauth==0.41.0
django-crispy-forms==1.8.1
django-markdownx==2.0.28
django-storages==1.8
docutils==0.15.2
ffmpeg-python==0.2.0
filetype==1.0.5
future==0.18.2
gunicorn==20.0.4
idna==2.8
jmespath==0.9.4
Markdown==3.1.1
mysqlclient==1.4.6
oauthlib==3.1.0
Pillow==6.2.1
psycopg2-binary==2.8.4
pyasn1==0.4.8
pycparser==2.19
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.2
requests==2.22.0
requests-oauthlib==1.3.0
rsa==3.4.2
s3transfer==0.3.1
six==1.13.0
soupsieve==1.9.5
sqlparse==0.3.0
urllib3==1.25.7
whitenoise==5.0.1
